refactor(client): split client.py into per-domain mixins
client.py at 1856 LOC put 44 methods on a single VerisureOwaClient class
spanning auth, alarm, activity, lock, camera, sentinel, and installation
operations. Convert to a client/ package with one mixin per domain plus
a shared _ClientBase that holds __init__, headers, GraphQL execute, and
the polling helpers:

- __init__.py — VerisureOwaClient (composes mixins) + module-level
  helpers (generate_uuid, generate_device_id) + re-exports
- _base.py — _ClientBase with shared state and infrastructure (620 LOC)
- _auth.py — login / refresh / logout / 2FA (230 LOC)
- _alarm.py — arm / disarm / check_alarm / status (270 LOC)
- _activity.py — panel timeline (40 LOC)
- _lock.py — smart-lock status / config / mode change (230 LOC)
- _camera.py — devices / capture / thumbnail / full image (270 LOC)
- _sentinel.py — comfort sensors and air-quality (120 LOC)
- _installation.py — installation list + service catalog (160 LOC)

Two no-member calls in _ClientBase (refresh_token, login, get_services)
that bind to mixin methods get explicit pylint disables since pylint
can't see them on the base class. Pure file-level reorganisation — no
behaviour change. All 1378 tests pass.
